moved existing line chart features to a utils file. added compact mode
